+++ packages/spotlight/CHANGELOG.md
@@ -8,6 +8,7 @@ The following is a curated list of changes in the Enact spotlight module, newest
 
 - `isMuted()` method to `@enact/spotlight`
 - `spotlightMuted` prop to `@enact/spotlight/container`
+- `isPaused()`, `isSpottable()`, and `getCurrent()` methods
 
 ## [1.0.0-alpha.5] - 2016-12-16